Fix graph visualization in colab
