+++ frontend/src/app.js
@@ -284,7 +284,6 @@ class SpannerApp {
                     <div id="sidebar-${this.id}"></div>
                     <div id="table-${this.id}" class="hidden"></div>
                 </div>
-                <p><b>Hello</b></p>
             </div>
         `;
